Added constraints
added constraints between each point_mass to for more realistic simulation of physics. Also changed the type of every float to double in order for more precision in calculation. Fixed a bug in the creation of a Body where one point_mass would be created on top of its origin point on the same side. Currently still having an issue with the parameter value of Young's modulus inside the Material struct. Right now i cant pinpoint if the parameter value should be in Pasclas, GPa, or MPa. Technically the calculations *should* are in Pa; however changing the value from GPa to Pa causes major instabilities.